* Renamed module to numexpr3 to allow it to be benchmarked relative to 2.5 * Built the block.hpp at compile time from Python to allow the block size to be optimized by the setup in the future * Fixed some Windows related bugs.  Windows is still not producing correct results, probably an endian issue, but it doesn't crash now.
