+++ MANIFEST.in
@@ -2,9 +2,9 @@ include MANIFEST.in
 
 include *.rst *.txt *.cfg site.cfg.example
 
-recursive-include numexpr *.cpp *.hpp *.py
-recursive-include numexpr/win32 *.c *.h
-exclude numexpr/__config__.py
+recursive-include numexpr3 *.cpp *.hpp *.py
+recursive-include numexpr3/win32 *.c *.h
+exclude numexpr3/__config__.py
 exclude RELEASING.txt
 
 recursive-include bench *.py
