+++ numexpr3/cpuinfo.py
