+++ numexpr3/missing_posix_functions.hpp
@@ -14,7 +14,7 @@
   like MSVC 7.1  */
 
 
-/* Double precision versions */
+// Double precision versions
 
 inline double log1p(double x)
 {
@@ -72,7 +72,7 @@ inline double atanh(double x)
 }
 
 
-/* Single precision versions */
+// Single precision versions 
 
 inline float log1pf(float x)
 {
@@ -97,6 +97,11 @@ inline float acoshf(float x)
 inline float atanhf(float x)
 {
     return (float) atanh((double)x);
+}
+
+inline float hypotf( float x, float y)
+{
+    return (float) hypot( (double)x, (double)y );
 }
 
 #endif // NUMEXPR_MISSING_POSIX_FUNCTIONS_HPP
