+++ numexpr3/numexpr_object.cpp
