+++ numexpr3/numexpr_object.hpp
@@ -1,0 +1,40 @@
+#ifndef NUMEXPR_OBJECT_HPP
+#define NUMEXPR_OBJECT_HPP
+/*********************************************************************
+  Numexpr - Fast numerical array expression evaluator for NumPy.
+
+      License: MIT
+      Author:  See AUTHORS.txt
+
+  See LICENSE.txt for details about copyright and rights to use.
+**********************************************************************/
+
+
+
+struct NumExprObject
+{
+    PyObject_HEAD
+    PyObject *signature;    /* a python string */
+    PyObject *tempsig;
+    PyObject *constsig;
+    //PyObject *fullsig;
+    PyObject *program_bytes;      // a python bytearray 
+    // RAM: words are arrays of ints converted from the byte arrays
+    unsigned short *program_words, *sig_words, *tempsig_words, *constsig_words; 
+    
+    PyObject *constants;    /* a tuple of int/float/complex */
+    PyObject *input_names;  /* tuple of strings */
+    char **mem;             /* pointers to registers */
+    char *rawmem;           /* a chunks of raw memory for storing registers */
+    npy_intp *memsteps;
+    npy_intp *memsizes;
+    int  rawmemsize;
+    int  n_inputs;
+    int  n_constants;
+    int  n_temps;
+    int word_count; // RAM: number of words in program_words
+};
+
+extern PyTypeObject NumExprType;
+
+#endif // NUMEXPR_OBJECT_HPP
