+++ numexpr3/tests/test_numexpr.py
