+++ numexpr3/version.py
@@ -8,4 +8,4 @@
 #  rights to use.
 ####################################################################
 
-version = '2.4.4.dev0'
+version = '3.0.0a0'
