+++ setup.py
@@ -16,24 +16,42 @@ import os.path as op
 from distutils.command.clean import clean
 import time
 
-# STARTmonkey-patch for parallel compilation (on Linux?)
-def parallelCCompile(self, sources, output_dir=None, macros=None, include_dirs=None, debug=0, extra_preargs=None, extra_postargs=None, depends=None):
-    # those lines are copied from distutils.ccompiler.CCompiler directly
-    macros, objects, extra_postargs, pp_opts, build = self._setup_compile(output_dir, macros, include_dirs, sources, depends, extra_postargs)
-    cc_args = self._get_cc_args(pp_opts, debug, extra_preargs)
-    # parallel code
-    N=8 # number of parallel compilations
-    import multiprocessing.pool
-    def _single_compile(obj):
-        try: src, ext = build[obj]
-        except KeyError: return
-        self._compile(obj, src, ext, cc_args, extra_postargs, pp_opts)
-    # convert to list, imap is evaluated on-demand
-    list(multiprocessing.pool.ThreadPool(N).imap(_single_compile,objects))
-    return objects
-import distutils.ccompiler
-distutils.ccompiler.CCompiler.compile=parallelCCompile
-# END monkey-patch for parrallel compilation
+"""
+ NOTES FOR WINDOWS:
+ For Python 2.7, you must have the x64 environment variables set correctly...
+
+from:
+    http://scikit-learn.org/stable/developers/advanced_installation.html#windows
+for the 64-bit architecture, you either need the full visual studio or the free windows sdks that 
+can be downloaded from the links below.
+the windows sdks include the msvc compilers both for 32 and 64-bit architectures. they come as a 
+grmsdkx_en_dvd.iso file that can be mounted as a new drive with a setup.exe installer in it.
+
+for python 2 you need sdk v7.0: ms windows sdk for windows 7 and .net framework 3.5 sp1
+
+for python 3 you need sdk v7.1: ms windows sdk for windows 7 and .net framework 4
+
+both sdks can be installed in parallel on the same host. to use the windows sdks, you need to setup 
+the environment of a cmd console launched with the following flags (at least for sdk v7.0):
+cmd /e:on /v:on /k
+then configure the build environment with (FOR PYTHON 2.7):
+
+set distutils_use_sdk=1
+set mssdk=1
+"c:\program files\microsoft sdks\windows\v7.0\setup\windowssdkver.exe" -q -version:v7.0
+"c:\program files\microsoft sdks\windows\v7.0\bin\setenv.cmd" /x64 /release
+
+finally you can build scikit-learn in the same cmd console:
+
+python setup.py install
+
+replace v7.0 by the v7.1 in the above commands to do the same for python 3 instead of python 2.
+replace /x64 by /x86 to build for 32-bit python instead of 64-bit python.
+
+Ignore the "Missing compiler_cxx fix for MSVCCompiler" error message.
+
+You also need the .NET Framework 3.5 SP1 installed for Python 2.7
+"""
 
 if sys.version_info < (2, 6):
     raise RuntimeError("must use python 2.6 or greater")
@@ -47,8 +65,39 @@ with open('requirements.txt') as f:
     requirements = f.read().splitlines()
 
 # Fetch the version for numexpr (will be put in variable `version`)
-with open(os.path.join('numexpr', 'version.py')) as f:
+with open(os.path.join('numexpr3', 'version.py')) as f:
     exec(f.read())
+    
+    
+# RAM: NEW compile-time definition of #define BLOCK_SIZE variables
+blocks_text = """/*********************************************************************
+  Numexpr - Fast numerical array expression evaluator for NumPy.
+
+      License: MIT
+      Author:  See AUTHORS.txt
+
+  See LICENSE.txt for details about copyright and rights to use.
+**********************************************************************/
+
+#ifdef USE_VML
+// The values below have been tuned for a nowadays Core2 processor 
+// Note: with VML functions a larger block size (e.g. 4096) allows to make use
+// of the automatic multithreading capabilities of the VML library 
+#define BLOCK_SIZE1 4096
+#define BLOCK_SIZE2 32
+#else
+// The values below have been tuned for a nowadays Core2 processor
+// Note: without VML available a smaller block size is best, specially
+// for the strided and unaligned cases.  Recent implementation of
+// multithreading make it clear that larger block sizes benefit
+// performance (although it seems like we don't need very large sizes
+// like VML yet). 
+#define BLOCK_SIZE1 1024
+#define BLOCK_SIZE2 16
+#endif
+"""
+with open( "numexpr3/blocks.hpp", 'wb' ) as fh:
+     fh.writelines( blocks_text )
 
 def setup_package():
     metadata = dict(
@@ -57,7 +106,7 @@ def setup_package():
                       author_email='david.m.cooke@gmail.com, faltet@gmail.com',
                       url='https://github.com/pydata/numexpr',
                       license='MIT',
-                      packages=['numexpr'],
+                      packages=['numexpr3'],
                       install_requires=requirements,
                       setup_requires=requirements
     )
@@ -75,7 +124,7 @@ def setup_package():
         except ImportError:
             from distutils.core import setup
 
-        metadata['name']    = 'numexpr'
+        metadata['name']    = 'numexpr3'
         metadata['version'] = version
     else:
         from numpy.distutils.core import setup
@@ -130,7 +179,7 @@ def setup_package():
             from numpy.distutils.misc_util import Configuration, dict_append
             from numpy.distutils.system_info import system_info
 
-            config = Configuration('numexpr')
+            config = Configuration('numexpr3')
 
             #try to find configuration for MKL, either from environment or site.cfg
             if op.exists('site.cfg'):
@@ -146,21 +195,21 @@ def setup_package():
 
             #setup information for C extension
             if os.name == 'nt':
-                pthread_win = ['numexpr/win32/pthread.c']
+                pthread_win = ['numexpr3/win32/pthread.c']
             else:
                 pthread_win = []
             extension_config_data = {
-                'sources': ['numexpr/interpreter.cpp',
-                            'numexpr/module.cpp',
-                            'numexpr/numexpr_object.cpp'] + pthread_win,
-                'depends': ['numexpr/interp_body.cpp',
-                            'numexpr/complex_functions.hpp',
-                            'numexpr/interpreter.hpp',
-                            'numexpr/module.hpp',
-                            'numexpr/msvc_function_stubs.hpp',
-                            'numexpr/numexpr_config.hpp',
-                            'numexpr/numexpr_object.hpp',
-			    'numexpr/opcodes.hpp'],
+                'sources': ['numexpr3/interpreter.cpp',
+                            'numexpr3/module.cpp',
+                            'numexpr3/numexpr_object.cpp'] + pthread_win,
+                'depends': ['numexpr3/interp_body.cpp',
+                            'numexpr3/complex_functions.hpp',
+                            'numexpr3/interpreter.hpp',
+                            'numexpr3/module.hpp',
+                            'numexpr3/msvc_function_stubs.hpp',
+                            'numexpr3/numexpr_config.hpp',
+                            'numexpr3/numexpr_object.hpp',
+                            'numexpr3/opcodes.hpp'],
                 'libraries': ['m'],
                 'extra_compile_args': ['-funroll-all-loops', ],
             }
@@ -170,10 +219,10 @@ def setup_package():
             config.add_extension('interpreter', **extension_config_data)
 
             config.make_config_py()
-            config.add_subpackage('tests', 'numexpr/tests')
+            config.add_subpackage('tests', 'numexpr3/tests')
 
             #version handling
-            config.get_version('numexpr/version.py')
+            config.get_version('numexpr3/version.py')
             return config
 
 
@@ -202,9 +251,9 @@ def setup_package():
                     for suffix, _, _ in imp.get_suffixes():
                         if suffix == '.py':
                             continue
-                        paths.append(localpath("numexpr", "interpreter" + suffix))
-                paths.append(localpath("numexpr/__config__.py"))
-                paths.append(localpath("numexpr/__config__.pyc"))
+                        paths.append(localpath("numexpr3", "interpreter" + suffix))
+                paths.append(localpath("numexpr3/__config__.py"))
+                paths.append(localpath("numexpr3/__config__.pyc"))
                 for path in paths:
                     try:
                         os.remove(path)
@@ -243,4 +292,4 @@ if __name__ == '__main__':
     t0 = time.time()
     setup_package()
     t1 = time.time()
-    print( "error: build/install time (s): " + str(t1-t0) )
+    print( "No error on " + sys.argv[1] +  " in time (s): " + str(t1-t0) )
